requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
clip==1.0
colorama==0.4.4
cycler==0.11.0
Cython==0.29.26
cytoolz==0.11.2
dnspython==2.1.0
einops==0.4.0
email-validator==1.1.3
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
filelock==3.4.2
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.29.0
frozenlist==1.3.0
fsspec==2022.1.0
ftfy==6.0.3
future==0.18.2
google-auth==2.4.1
google-auth-oauthlib==0.4.6
greenlet==1.1.2
grpcio==1.43.0
gunicorn==20.1.0
hexbytes==0.2.2
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.10.1
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
kiwisolver==1.3.2
lru-dict==1.1.7
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.22.1
oauthlib==3.1.1
omegaconf==2.1.1
packaging==21.3
parsimonious==0.8.1
Pillow==9.0.0
protobuf==3.19.3
py-geth==3.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.1
pyDeprecate==0.3.1
pyflakes==2.4.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-lightning==1.5.9
pywin32==303
PyYAML==6.0
q==2.6
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.0
rlp==2.0.1
rsa==4.8
sacremoses==0.0.47
semantic-version==2.8.5
six==1.16.0
SQLAlchemy==1.4.29
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tokenizers==0.10.3
toolz==0.11.2
torch==1.8.0
torchmetrics==0.7.0
torchvision==0.9.0
tqdm==4.62.3
transformers==4.15.0
typing_extensions==4.0.1
urllib3==1.26.8
varint==1.0.2
wcwidth==0.2.5
web3==5.28.0
websockets==9.1
Werkzeug==2.0.2
WTForms==3.0.1
yarl==1.7.2
zipp==3.7.0
